@@ -1,26 +1,19 @@
 {
   "name": "vue-markdown-edit-example",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
+  "version": "0.0.0",
   "scripts": {
-    "serve": "rollup -cw"
+    "dev": "vite",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "punycode": "^2.1.1",
-    "vue": "^2.6.12",
-    "vue-markdown-render": "^1.0.2"
+    "vue": "^3.2.25",
+    "vue-markdown-render": "^2.0.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^15.1.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^9.0.0",
-    "@rollup/plugin-replace": "^2.3.3",
-    "@rollup/plugin-typescript": "^6.0.0",
-    "rollup": "^2.31.0",
-    "rollup-plugin-css-only": "^2.1.0",
-    "rollup-plugin-serve": "^1.0.4",
-    "rollup-plugin-terser": "^7.0.2",
-    "tslib": "^2.0.3"
+    "@vitejs/plugin-vue": "^2.0.1",
+    "typescript": "^4.5.4",
+    "vite": "^2.7.8",
+    "vue-tsc": "^0.30.1"
   }
 }
