@@ -17,11 +17,12 @@
     "build": "tsc"
   },
   "dependencies": {
-    "codemirror": "^5.58.1",
-    "vue": "^2.6.12"
+    "codemirror": "^5.65.0",
+    "vue": "^3.2.25"
   },
   "devDependencies": {
-    "@types/codemirror": "^0.0.98",
-    "typescript": "^4.0.3"
+    "@babel/types": "^7.16.0",
+    "@types/codemirror": "^5.60.5",
+    "typescript": "^4.5.4"
   }
 }
