@@ -1,13 +1,13 @@
-import Vue, { VNode } from "vue";
-import createEditor, { Editor } from "codemirror";
-import "codemirror/mode/markdown/markdown";
-import "codemirror/lib/codemirror.css";
-import "../styles/MarkdownEditor.css";
+import { defineComponent, h, VNode } from 'vue'
+import createEditor, { Editor } from 'codemirror'
+import 'codemirror/mode/markdown/markdown'
+import 'codemirror/lib/codemirror.css'
+import '../styles/MarkdownEditor.css'
 
-const MarkdownEditor = Vue.extend({
-  name: "MarkdownEditor",
+const MarkdownEditor = defineComponent({
+  name: 'MarkdownEditor',
   props: {
-    value: {
+    modelValue: {
       type: String,
       required: true,
     },
@@ -15,28 +15,28 @@ const MarkdownEditor = Vue.extend({
   data() {
     return {
       editor: null as Editor | null,
-    };
+    }
   },
   mounted() {
     this.editor = createEditor(this.$el as HTMLElement, {
-      mode: "markdown",
+      mode: 'markdown',
       lineNumbers: true,
-      value: this.value,
-    });
+      value: this.modelValue,
+    })
 
-    this.editor.on("change", this.onChange);
+    this.editor.on('change', this.onChange)
   },
   beforeDestroy() {
-    this.editor?.off("change", this.onChange);
+    this.editor?.off('change', this.onChange)
   },
   methods: {
     onChange() {
-      this.$emit("input", this.editor?.getValue());
+      this.$emit('update:modelValue', this.editor?.getValue())
     },
   },
-  render(h): VNode {
-    return h("div", { class: "markdown-editor" });
+  render(): VNode {
+    return h('div', { class: 'markdown-editor' })
   },
-});
+})
 
-export default MarkdownEditor;
+export default MarkdownEditor
